backend/service_reports/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
httpx==0.27.2
openpyxl==3.1.5




